requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<4.0
drf-spectacular>=0.26.3

autoflake
autopep8
flake8